requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.17
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
environs==11.2.1
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
marshmallow==3.23.1
multidict==6.1.0
packaging==24.2
pillow==11.0.0
propcache==0.2.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
yarl==1.18.3
